# contract/openapi.yaml
openapi: 3.1.0
info:
  title: cmProjectX Service API
  version: 0.1.0
  description: >
    Contract between the Rust/WinUI 3 client and the .NET Intune engine.
    This is the single source of truth for DTOs and endpoints — generate
    C# and Rust types from it (see contract/README.md).
servers:
  - url: http://127.0.0.1:5099
    description: Local sidecar (MVP)
paths:
...
  /device-configs:
    get:
      operationId: listDeviceConfigs
      summary: Device configurations (live Graph; 409 when signed out)
      responses: { "200": { $ref: "#/components/responses/ListItemArray" }, "409": { description: Not signed in } }
    post:
      operationId: createDeviceConfig
      summary: Create a device configuration
      requestBody: { $ref: "#/components/requestBodies/ObjectJson" }
      responses: { "200": { $ref: "#/components/responses/Created" }, "409": { description: Not signed in } }
  /compliance-policies:
    get:
      operationId: listCompliancePolicies
      summary: Device compliance policies
      responses: { "200": { $ref: "#/components/responses/ListItemArray" }, "409": { description: Not signed in } }
    post:
      operationId: createCompliancePolicy
      summary: Create a compliance policy
      requestBody: { $ref: "#/components/requestBodies/ObjectJson" }
      responses: { "200": { $ref: "#/components/responses/Created" }, "409": { description: Not signed in } }
  /settings-catalog:
    get:
      operationId: listSettingsCatalog
      summary: Settings Catalog policies
      responses: { "200": { $ref: "#/components/responses/ListItemArray" }, "409": { description: Not signed in } }
    post:
      operationId: createSettingsCatalog
      summary: Create a Settings Catalog policy
      requestBody: { $ref: "#/components/requestBodies/ObjectJson" }
      responses: { "200": { $ref: "#/components/responses/Created" }, "409": { description: Not signed in } }
  /admin-templates:
    get:
      operationId: listAdminTemplates
      summary: Administrative templates (group policy configurations)
      responses: { "200": { $ref: "#/components/responses/ListItemArray" }, "409": { description: Not signed in } }
    post:
      operationId: createAdminTemplate
      summary: Create an administrative template
      requestBody: { $ref: "#/components/requestBodies/ObjectJson" }
      responses: { "200": { $ref: "#/components/responses/Created" }, "409": { description: Not signed in } }
  /endpoint-security:
    get:
      operationId: listEndpointSecurity
      summary: Endpoint security intents
      responses: { "200": { $ref: "#/components/responses/ListItemArray" }, "409": { description: Not signed in } }
    post:
      operationId: createEndpointSecurity
      summary: Create an endpoint security intent
      requestBody: { $ref: "#/components/requestBodies/ObjectJson" }
      responses: { "200": { $ref: "#/components/responses/Created" }, "409": { description: Not signed in } }
...
  /device-configs/{id}:
    get:
      operationId: getDeviceConfig
      summary: Device configuration object JSON
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      responses: { "200": { $ref: "#/components/responses/ObjectJson" }, "404": { description: Not found }, "409": { description: Not signed in } }
    patch:
      operationId: updateDeviceConfig
      summary: Update a device configuration
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      requestBody: { $ref: "#/components/requestBodies/ObjectJson" }
      responses: { "204": { description: Updated }, "409": { description: Not signed in } }
    delete:
      operationId: deleteDeviceConfig
      summary: Delete a device configuration
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      responses: { "204": { description: Deleted }, "409": { description: Not signed in } }
  /compliance-policies/{id}:
    get:
      operationId: getCompliancePolicy
      summary: Compliance policy object JSON
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      responses: { "200": { $ref: "#/components/responses/ObjectJson" }, "404": { description: Not found }, "409": { description: Not signed in } }
    patch:
      operationId: updateCompliancePolicy
      summary: Update a compliance policy
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      requestBody: { $ref: "#/components/requestBodies/ObjectJson" }
      responses: { "204": { description: Updated }, "409": { description: Not signed in } }
    delete:
      operationId: deleteCompliancePolicy
      summary: Delete a compliance policy
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      responses: { "204": { description: Deleted }, "409": { description: Not signed in } }
  /settings-catalog/{id}:
    get:
      operationId: getSettingsCatalog
      summary: Settings Catalog policy object JSON
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      responses: { "200": { $ref: "#/components/responses/ObjectJson" }, "404": { description: Not found }, "409": { description: Not signed in } }
    patch:
      operationId: updateSettingsCatalog
      summary: Update a Settings Catalog policy (metadata)
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      requestBody: { $ref: "#/components/requestBodies/ObjectJson" }
      responses: { "204": { description: Updated }, "409": { description: Not signed in } }
    delete:
      operationId: deleteSettingsCatalog
      summary: Delete a Settings Catalog policy
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      responses: { "204": { description: Deleted }, "409": { description: Not signed in } }
  /settings-catalog/{id}/settings:
    get:
      operationId: getSettingsCatalogSettings
      summary: The deep settingInstance tree for a Settings Catalog policy (fetched apart from metadata)
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      responses:
        "200":
          description: The policy's settings as a JSON array
          content:
            application/json:
              schema: { type: array, items: { type: object } }
        "409": { description: Not signed in }
  /admin-templates/{id}:
    get:
      operationId: getAdminTemplate
      summary: Administrative template object JSON
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      responses: { "200": { $ref: "#/components/responses/ObjectJson" }, "404": { description: Not found }, "409": { description: Not signed in } }
    patch:
      operationId: updateAdminTemplate
      summary: Update an administrative template
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      requestBody: { $ref: "#/components/requestBodies/ObjectJson" }
      responses: { "204": { description: Updated }, "409": { description: Not signed in } }
    delete:
      operationId: deleteAdminTemplate
      summary: Delete an administrative template
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      responses: { "204": { description: Deleted }, "409": { description: Not signed in } }
  /endpoint-security/{id}:
    get:
      operationId: getEndpointSecurity
      summary: Endpoint security intent object JSON
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      responses: { "200": { $ref: "#/components/responses/ObjectJson" }, "404": { description: Not found }, "409": { description: Not signed in } }
    patch:
      operationId: updateEndpointSecurity
      summary: Update an endpoint security intent
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      requestBody: { $ref: "#/components/requestBodies/ObjectJson" }
      responses: { "204": { description: Updated }, "409": { description: Not signed in } }
    delete:
      operationId: deleteEndpointSecurity
      summary: Delete an endpoint security intent
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      responses: { "204": { description: Deleted }, "409": { description: Not signed in } }
...
  /device-configs/{id}/assignments:
    get:
      operationId: getDeviceConfigAssignments
      summary: Normalized assignment set for a device configuration (read-only surface, no replace)
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      responses:
        "200": { description: Current assignments, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Assignment" } } } } }
        "409": { description: Not signed in }
  /compliance-policies/{id}/assignments:
    get:
      operationId: getComplianceAssignments
      summary: Normalized assignment set (group names resolved best-effort)
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      responses:
        "200": { description: Current assignments, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Assignment" } } } } }
        "409": { description: Not signed in }
    post:
      operationId: setComplianceAssignments
      summary: Replace ALL assignments on a compliance policy (replace-all write)
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      requestBody: { required: true, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Assignment" } } } } }
      responses: { "204": { description: Assignments replaced }, "409": { description: Not signed in } }
  /admin-templates/{id}/assignments:
    get:
      operationId: getAdminTemplateAssignments
      summary: Normalized assignment set for an administrative template
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      responses:
        "200": { description: Current assignments, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Assignment" } } } } }
        "409": { description: Not signed in }
    post:
      operationId: setAdminTemplateAssignments
      summary: Replace ALL assignments on an administrative template (replace-all write)
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      requestBody: { required: true, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Assignment" } } } } }
      responses: { "204": { description: Assignments replaced }, "409": { description: Not signed in } }
...
components:
  responses:
    ListItemArray:
      description: Normalized list rows for the active tenant
      content:
        application/json:
          schema:
            type: array
            items: { $ref: "#/components/schemas/ListItem" }
    ObjectJson:
      description: The full resource as a free-form Graph object (normalized JSON)
      content:
        application/json:
          schema: { type: object, additionalProperties: true }
    Created:
      description: Created; returns the new object's id
      content:
        application/json:
          schema:
            type: object
            properties: { id: { type: string, nullable: true } }
  requestBodies:
    ObjectJson:
      description: The resource as a free-form Graph object (normalized JSON); the @odata.type drives deserialization
      required: true
      content:
        application/json:
          schema: { type: object, additionalProperties: true }
  schemas:
...
    Assignment:
      type: object
      description: >
        Normalized Intune assignment row shared by every assignable surface
        (mirrors api_types::Assignment / AssignmentDto byte-for-byte). The sidecar
        translates this to/from the per-resource typed Graph assignment target.
        `intent` is apps-only (null elsewhere).
      properties:
        kind:       { type: string, description: "group | exclusionGroup | allDevices | allUsers" }
        groupId:    { type: string, nullable: true }
        groupName:  { type: string, nullable: true }
        filterId:   { type: string, nullable: true }
        filterMode: { type: string, nullable: true, description: "include | exclude" }
        intent:     { type: string, nullable: true, description: "apps: required | available | uninstall | availableWithoutEnrollment" }
      required: [kind]
...